fix : make drug data only server run
- manage.py 를 수행할 때마다 마약 데이터가 생성,삭제됨
- manage.py runserver 시에만 생성/삭제 되도록 수정

+++ apps/data_api/apps.py
@@ -1,11 +1,12 @@
 import logging
-from sqlite3 import OperationalError
-from django.db.utils import OperationalError
+import sys
+
+from django.db import OperationalError
 
 from django.apps import AppConfig
 from django.core.management import call_command
 
-logger = logging.getLogger(__name__)
+from apps.data_api.constants import DRUG_TABLE_NOT_EXIST
 
 
 class DataApiConfig(AppConfig):
@@ -13,13 +14,22 @@ class DataApiConfig(AppConfig):
     name = 'apps.data_api'
 
     def ready(self):
-        # ready 실행시 Drug 모델이 생성안될 수 있어 내부에서 호출
-        from .models import Drug
+        """
+         runserver시 drug data를 load 한다
+        """
+
+        # only run during 'runserver' command
+        if 'runserver' not in sys.argv:
+            return
 
         try:
+            # ready 실행시 Drug 모델이 생성이 안될 수 있어 함수 내부에서 호출
+            from .models import Drug
+
             if Drug.objects.exists():
                 Drug.objects.all().delete()
                 
             call_command('loaddata', 'drug.json')
         except OperationalError:
-            logger.warning("[drugs 데이터 생성 실패] drugs 테이블이 없습니다 migrate를 수행해주세요")
+            logging.warning(DRUG_TABLE_NOT_EXIST)
+            raise OperationalError(DRUG_TABLE_NOT_EXIST)
